refactor(cpu): separate ordinary and fused CPU kernel owners
