+++ crates/tenferro-ad/tests/ui/eager_backend_owner_private.stderr
@@ -2,9 +2,7 @@ error[E0432]: unresolved import `tenferro_ad::EagerBackend`
  --> tests/ui/eager_backend_owner_private.rs:1:5
   |
 1 | use tenferro_ad::EagerBackend;
-  |     ^^^^^^^^^^^^^------------
-  |                  |
-  |                  no `EagerBackend` in the root
+  |     ^^^^^^^^^^^^^^^^^^^^^^^^^ no `EagerBackend` in the root
   |
 help: a similar name exists in the module
   |
