+++ crates/tenferro-cpu-basic/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "tenferro-internal-cpu-kernels"
+name = "tenferro-cpu-basic"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
@@ -8,24 +8,22 @@ publish.workspace = true
 readme.workspace = true
 repository.workspace = true
 homepage.workspace = true
-description = "Internal ordinary CPU kernel implementations for tenferro-cpu."
+description = "Shared CPU execution resources and host strided helpers for tenferro."
 rust-version.workspace = true
 keywords = ["tensor", "kernels", "cpu", "numerical"]
 categories = ["science", "mathematics", "hardware-support"]
-documentation = "https://docs.rs/tenferro-internal-cpu-kernels"
+documentation = "https://docs.rs/tenferro-cpu-basic"
 
 [package.metadata.docs.rs]
 rustdoc-args = ["--cfg", "docsrs"]
 all-features = true
 
 [lib]
-name = "tenferro_internal_cpu_kernels"
+name = "tenferro_cpu_basic"
 
 [dependencies]
-tenferro-cpu-basic = { path = "../tenferro-cpu-basic", version = "0.5.0" }
 tenferro-tensor = { path = "../tenferro-tensor", version = "0.5.0" }
 num-complex.workspace = true
 num-traits.workspace = true
-smallvec.workspace = true
-strided-kernel.workspace = true
+strided-basic.workspace = true
 thiserror.workspace = true
