+++ crates/tenferro-cpu-basic/src/pooled_uninit_output/tests.rs
