+++ crates/tenferro-cpu-fused/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "tenferro-internal-cpu-kernels"
+name = "tenferro-cpu-fused"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
@@ -8,24 +8,27 @@ publish.workspace = true
 readme.workspace = true
 repository.workspace = true
 homepage.workspace = true
-description = "Internal ordinary CPU kernel implementations for tenferro-cpu."
+description = "Fused CPU elementwise execution for tenferro."
 rust-version.workspace = true
 keywords = ["tensor", "kernels", "cpu", "numerical"]
 categories = ["science", "mathematics", "hardware-support"]
-documentation = "https://docs.rs/tenferro-internal-cpu-kernels"
+documentation = "https://docs.rs/tenferro-cpu-fused"
 
 [package.metadata.docs.rs]
 rustdoc-args = ["--cfg", "docsrs"]
 all-features = true
 
 [lib]
-name = "tenferro_internal_cpu_kernels"
+name = "tenferro_cpu_fused"
 
 [dependencies]
 tenferro-cpu-basic = { path = "../tenferro-cpu-basic", version = "0.5.0" }
 tenferro-tensor = { path = "../tenferro-tensor", version = "0.5.0" }
 num-complex.workspace = true
 num-traits.workspace = true
-smallvec.workspace = true
-strided-kernel.workspace = true
+strided-basic.workspace = true
+strided-fused.workspace = true
 thiserror.workspace = true
+
+[dev-dependencies]
+criterion.workspace = true
