+++ crates/tenferro-tensor/Cargo.toml
@@ -29,7 +29,6 @@ tenferro-tensor-core = { path = "../tenferro-tensor-core", version = "0.5.0" }
 num-traits.workspace = true
 num-complex.workspace = true
 smallvec.workspace = true
-strided-kernel.workspace = true
 tenferro-core-ops = { path = "../tenferro-core-ops", version = "0.5.0" }
 thiserror.workspace = true
