+++ crates/tenferro-tensor/src/backend.rs
@@ -11,13 +11,7 @@ use crate::{
     TensorRead, TensorValue, TensorWrite, ValidationError,
 };
 use num_complex::{Complex32, Complex64};
-use smallvec::SmallVec;
 use std::any::TypeId;
-use std::ptr::NonNull;
-use strided_kernel::{
-    erased_map_into, erased_zip_into, ErasedMapOp, ErasedRawStridedMut, ErasedRawStridedPtr,
-    ErasedZipOp, ExecContext, KernelDType,
-};
 
 #[cfg(test)]
 mod tests;
@@ -1492,7 +1486,8 @@ pub enum ElementwiseReadOp {
 }
 
 impl ElementwiseReadOp {
-    fn label(self) -> &'static str {
+    #[doc(hidden)]
+    pub fn label(self) -> &'static str {
         match self {
             Self::Add => "add",
             Self::Subtract => "sub",
@@ -1503,7 +1498,8 @@ impl ElementwiseReadOp {
         }
     }
 
-    fn arity(self) -> usize {
+    #[doc(hidden)]
+    pub fn arity(self) -> usize {
         match self {
             Self::Negate | Self::Conj => 1,
             Self::Add | Self::Subtract | Self::Multiply | Self::Divide => 2,
@@ -1644,14 +1640,6 @@ fn storage_overlaps(lhs: StorageIdentity, rhs: StorageIdentity) -> bool {
     }
 }
 
-fn validate_elementwise_output_disjoint(
-    op: ElementwiseReadOp,
-    inputs: &[TensorRead<'_>],
-    out: &TensorWrite<'_>,
-) -> crate::Result<()> {
-    validate_read_into_destination(op.label(), inputs, out)
-}
-
 /// Validate that a caller-owned destination does not overlap any read input.
 ///
 /// The check is intentionally conservative for host views: two views backed by
@@ -1699,307 +1687,22 @@ pub fn validate_read_into_destination(
     Ok(())
 }
 
-fn read_is_host(input: &TensorRead<'_>) -> bool {
-    match input {
-        TensorRead::Tensor(tensor) => !tensor.is_backend_buffer(),
-        TensorRead::View(view) => match view {
-            TensorView::F32(view) => view.backend_buffer().is_none(),
-            TensorView::F64(view) => view.backend_buffer().is_none(),
-            TensorView::I32(view) => view.backend_buffer().is_none(),
-            TensorView::I64(view) => view.backend_buffer().is_none(),
-            TensorView::Bool(view) => view.backend_buffer().is_none(),
-            TensorView::C32(view) => view.backend_buffer().is_none(),
-            TensorView::C64(view) => view.backend_buffer().is_none(),
-        },
-    }
-}
-
-fn write_is_host(out: &TensorWrite<'_>) -> bool {
-    read_is_host(&out.as_read())
-}
-
-fn one_shot_supports(op: ElementwiseReadOp, dtype: DType) -> bool {
-    match op {
-        ElementwiseReadOp::Conj => true,
-        ElementwiseReadOp::Add
-        | ElementwiseReadOp::Subtract
-        | ElementwiseReadOp::Multiply
-        | ElementwiseReadOp::Divide
-        | ElementwiseReadOp::Negate => !matches!(dtype, DType::Bool),
-    }
-}
-
-fn one_shot_eligible(
-    op: ElementwiseReadOp,
-    inputs: &[TensorRead<'_>],
-    out: &TensorWrite<'_>,
-) -> bool {
-    let dtype = out.dtype();
-    write_is_host(out)
-        && one_shot_supports(op, dtype)
-        && inputs.iter().all(|input| {
-            read_is_host(input) && input.dtype() == dtype && input.shape() == out.shape()
-        })
-}
-
-fn tensor_write_view(out: TensorWrite<'_>) -> TensorViewMut<'_> {
-    match out {
-        TensorWrite::Tensor(tensor) => match tensor {
-            Tensor::F32(tensor) => TensorViewMut::F32(tensor.as_view_mut()),
-            Tensor::F64(tensor) => TensorViewMut::F64(tensor.as_view_mut()),
-            Tensor::I32(tensor) => TensorViewMut::I32(tensor.as_view_mut()),
-            Tensor::I64(tensor) => TensorViewMut::I64(tensor.as_view_mut()),
-            Tensor::Bool(tensor) => TensorViewMut::Bool(tensor.as_view_mut()),
-            Tensor::C32(tensor) => TensorViewMut::C32(tensor.as_view_mut()),
-            Tensor::C64(tensor) => TensorViewMut::C64(tensor.as_view_mut()),
-        },
-        TensorWrite::View(view) => view,
-    }
-}
-
-fn non_null_bytes<T>(data: &[T]) -> NonNull<u8> {
-    NonNull::new(data.as_ptr().cast_mut().cast()).unwrap_or_else(NonNull::dangling)
-}
-
-fn typed_bytes<T>(data: &[T]) -> &[u8] {
-    // SAFETY: u8 has alignment one and the returned bytes retain the shared
-    // lifetime of the typed source slice.
-    unsafe { std::slice::from_raw_parts(data.as_ptr().cast(), std::mem::size_of_val(data)) }
-}
-
-fn typed_bytes_mut<T>(data: &mut [T]) -> &mut [u8] {
-    let len = std::mem::size_of_val(data);
-    // SAFETY: u8 has alignment one and the returned bytes retain the unique
-    // lifetime of the typed destination slice.
-    unsafe { std::slice::from_raw_parts_mut(data.as_mut_ptr().cast(), len) }
-}
-
-fn erased_raw_strided_ptr<'a>(
-    dtype: KernelDType,
-    data: &'a [u8],
-    dims: &'a [usize],
-    strides: &'a [isize],
-    offset: isize,
-) -> strided_kernel::Result<ErasedRawStridedPtr<'a>> {
-    // SAFETY: callers derive `data` from initialized typed host storage and
-    // retain the backing borrow for the returned descriptor lifetime.
-    unsafe {
-        ErasedRawStridedPtr::from_raw_parts(
-            dtype,
-            non_null_bytes(data),
-            data.len(),
-            dims,
-            strides,
-            offset,
-        )
-    }
-}
-
-fn erased_raw_strided_mut<'a>(
-    dtype: KernelDType,
-    data: &'a mut [u8],
-    dims: &'a [usize],
-    strides: &'a [isize],
-    offset: isize,
-) -> strided_kernel::Result<ErasedRawStridedMut<'a>> {
-    let data_ptr = NonNull::new(data.as_mut_ptr()).unwrap_or_else(NonNull::dangling);
-    // SAFETY: callers derive `data` from a uniquely borrowed initialized host
-    // destination and retain that borrow for the returned descriptor lifetime.
-    unsafe {
-        ErasedRawStridedMut::from_raw_parts(dtype, data_ptr, data.len(), dims, strides, offset)
-    }
-}
-
-fn execute_one_shot_map<T: 'static>(
-    dtype: KernelDType,
-    op: ErasedMapOp,
-    ctx: &ExecContext,
-    input: TypedTensorView<'_, T>,
-    mut out: TypedTensorViewMut<'_, T>,
-) -> crate::Result<()> {
-    let input_data = input.host_storage()?;
-    // INVARIANT: dtype and layout come from the same validated typed view, and
-    // its host storage remains borrowed until replay returns.
-    // SAFETY: input_data supplies the pointer and exact byte length; the view
-    // owns the matching shape, signed strides, and in-bounds offset.
-    let input_descriptor = erased_raw_strided_ptr(
-        dtype,
-        typed_bytes(input_data),
-        input.shape(),
-        input.strides(),
-        input.offset(),
-    )
-    .map_err(|error| Error::backend_source("elementwise_read_into", error))?;
-
-    let out_dims = SmallVec::<[usize; 8]>::from_slice(out.shape());
-    let out_strides = SmallVec::<[isize; 8]>::from_slice(out.strides());
-    let out_offset = out.offset();
-    let out_data = out.host_storage_mut()?;
-    // INVARIANT: the copied output layout describes this uniquely borrowed
-    // host storage, already validated as disjoint from every input.
-    let mut out_descriptor = erased_raw_strided_mut(
-        dtype,
-        typed_bytes_mut(out_data),
-        &out_dims,
-        &out_strides,
-        out_offset,
-    )
-    .map_err(|error| Error::backend_source("elementwise_read_into", error))?;
-    erased_map_into(dtype, op, ctx, &mut out_descriptor, &input_descriptor)
-        .map_err(|error| Error::backend_source("elementwise_read_into", error))
-}
-
-fn execute_one_shot_zip<T: 'static>(
-    dtype: KernelDType,
-    op: ErasedZipOp,
-    ctx: &ExecContext,
-    lhs: TypedTensorView<'_, T>,
-    rhs: TypedTensorView<'_, T>,
-    mut out: TypedTensorViewMut<'_, T>,
-) -> crate::Result<()> {
-    let lhs_data = lhs.host_storage()?;
-    // INVARIANT: dtype and layout come from the same validated typed view, and
-    // its host storage remains borrowed until replay returns.
-    // SAFETY: lhs_data supplies the pointer and exact byte length; the view
-    // owns the matching shape, signed strides, and in-bounds offset.
-    let lhs_descriptor = erased_raw_strided_ptr(
-        dtype,
-        typed_bytes(lhs_data),
-        lhs.shape(),
-        lhs.strides(),
-        lhs.offset(),
-    )
-    .map_err(|error| Error::backend_source("elementwise_read_into", error))?;
-    let rhs_data = rhs.host_storage()?;
-    // INVARIANT: dtype and layout come from the same validated typed view, and
-    // its host storage remains borrowed until replay returns.
-    // SAFETY: rhs_data supplies the pointer and exact byte length; the view
-    // owns the matching shape, signed strides, and in-bounds offset.
-    let rhs_descriptor = erased_raw_strided_ptr(
-        dtype,
-        typed_bytes(rhs_data),
-        rhs.shape(),
-        rhs.strides(),
-        rhs.offset(),
-    )
-    .map_err(|error| Error::backend_source("elementwise_read_into", error))?;
-
-    let out_dims = SmallVec::<[usize; 8]>::from_slice(out.shape());
-    let out_strides = SmallVec::<[isize; 8]>::from_slice(out.strides());
-    let out_offset = out.offset();
-    let out_data = out.host_storage_mut()?;
-    // INVARIANT: the copied output layout describes this uniquely borrowed
-    // host storage, already validated as disjoint from every input.
-    let mut out_descriptor = erased_raw_strided_mut(
-        dtype,
-        typed_bytes_mut(out_data),
-        &out_dims,
-        &out_strides,
-        out_offset,
-    )
-    .map_err(|error| Error::backend_source("elementwise_read_into", error))?;
-    erased_zip_into(
-        dtype,
-        op,
-        ctx,
-        &mut out_descriptor,
-        &lhs_descriptor,
-        &rhs_descriptor,
-    )
-    .map_err(|error| Error::backend_source("elementwise_read_into", error))
-}
-
-fn execute_one_shot_elementwise(
-    op: ElementwiseReadOp,
-    inputs: &[TensorRead<'_>],
-    out: TensorWrite<'_>,
-    ctx: &ExecContext,
-) -> crate::Result<()> {
-    let out = tensor_write_view(out);
-    macro_rules! dispatch_map {
-        ($map_op:expr) => {{
-            let input = inputs[0].clone().tensor_view();
-            match (input, out) {
-                (TensorView::F32(input), TensorViewMut::F32(out)) => {
-                    execute_one_shot_map(KernelDType::F32, $map_op, ctx, input, out)
-                }
-                (TensorView::F64(input), TensorViewMut::F64(out)) => {
-                    execute_one_shot_map(KernelDType::F64, $map_op, ctx, input, out)
-                }
-                (TensorView::I32(input), TensorViewMut::I32(out)) => {
-                    execute_one_shot_map(KernelDType::I32, $map_op, ctx, input, out)
-                }
-                (TensorView::I64(input), TensorViewMut::I64(out)) => {
-                    execute_one_shot_map(KernelDType::I64, $map_op, ctx, input, out)
-                }
-                (TensorView::Bool(input), TensorViewMut::Bool(out)) => {
-                    execute_one_shot_map(KernelDType::Bool, $map_op, ctx, input, out)
-                }
-                (TensorView::C32(input), TensorViewMut::C32(out)) => {
-                    execute_one_shot_map(KernelDType::C32, $map_op, ctx, input, out)
-                }
-                (TensorView::C64(input), TensorViewMut::C64(out)) => {
-                    execute_one_shot_map(KernelDType::C64, $map_op, ctx, input, out)
-                }
-                _ => unreachable!("one-shot eligibility validates matching dtypes"),
-            }
-        }};
-    }
-    macro_rules! dispatch_zip {
-        ($zip_op:expr) => {{
-            let lhs = inputs[0].clone().tensor_view();
-            let rhs = inputs[1].clone().tensor_view();
-            match (lhs, rhs, out) {
-                (TensorView::F32(lhs), TensorView::F32(rhs), TensorViewMut::F32(out)) => {
-                    execute_one_shot_zip(KernelDType::F32, $zip_op, ctx, lhs, rhs, out)
-                }
-                (TensorView::F64(lhs), TensorView::F64(rhs), TensorViewMut::F64(out)) => {
-                    execute_one_shot_zip(KernelDType::F64, $zip_op, ctx, lhs, rhs, out)
-                }
-                (TensorView::I32(lhs), TensorView::I32(rhs), TensorViewMut::I32(out)) => {
-                    execute_one_shot_zip(KernelDType::I32, $zip_op, ctx, lhs, rhs, out)
-                }
-                (TensorView::I64(lhs), TensorView::I64(rhs), TensorViewMut::I64(out)) => {
-                    execute_one_shot_zip(KernelDType::I64, $zip_op, ctx, lhs, rhs, out)
-                }
-                (TensorView::C32(lhs), TensorView::C32(rhs), TensorViewMut::C32(out)) => {
-                    execute_one_shot_zip(KernelDType::C32, $zip_op, ctx, lhs, rhs, out)
-                }
-                (TensorView::C64(lhs), TensorView::C64(rhs), TensorViewMut::C64(out)) => {
-                    execute_one_shot_zip(KernelDType::C64, $zip_op, ctx, lhs, rhs, out)
-                }
-                _ => unreachable!("one-shot eligibility validates matching dtypes"),
-            }
-        }};
-    }
-
-    match op {
-        ElementwiseReadOp::Add => dispatch_zip!(ErasedZipOp::Add),
-        ElementwiseReadOp::Subtract => dispatch_zip!(ErasedZipOp::Subtract),
-        ElementwiseReadOp::Multiply => dispatch_zip!(ErasedZipOp::Multiply),
-        ElementwiseReadOp::Divide => dispatch_zip!(ErasedZipOp::Divide),
-        ElementwiseReadOp::Negate => dispatch_map!(ErasedMapOp::Negate),
-        ElementwiseReadOp::Conj => dispatch_map!(ErasedMapOp::Conj),
-    }
-}
-
-/// Execute the shared elementwise-into path with an explicit replay context.
+/// Execute a read-into elementwise operation through a backend's allocating operations.
 ///
-/// This is backend glue for implementations that own an execution context.
+/// Backend implementations use this helper when their device-native operation
+/// APIs already provide the correct placement and copy semantics. It deliberately
+/// does not inspect or transfer storage across devices.
 ///
 /// # Errors
 ///
-/// Returns [`crate::Error::Validation`] when the input arity or tensor
-/// metadata is invalid, or when the destination overlaps an input. Returns
-/// [`crate::Error::BackendSource`] when an eligible strided replay fails.
-/// Errors returned by `fallback` are preserved unchanged.
+/// Returns [`Error::Validation`] for wrong arity or overlapping output storage,
+/// and propagates the selected backend operation and copy errors.
 #[doc(hidden)]
-pub fn elementwise_read_into_with_context(
+pub fn elementwise_read_into_via_allocating_ops<B: TensorElementwise + ?Sized>(
+    backend: &mut B,
     op: ElementwiseReadOp,
     inputs: &[TensorRead<'_>],
     out: TensorWrite<'_>,
-    ctx: &ExecContext,
-    fallback: impl FnOnce(&[TensorRead<'_>], TensorWrite<'_>) -> crate::Result<()>,
 ) -> crate::Result<()> {
     if inputs.len() != op.arity() {
         return Err(Error::invalid_argument(
@@ -2008,12 +1711,16 @@ pub fn elementwise_read_into_with_context(
             format!("expected {} inputs, got {}", op.arity(), inputs.len()),
         ));
     }
-    validate_elementwise_output_disjoint(op, inputs, &out)?;
-    if one_shot_eligible(op, inputs, &out) {
-        execute_one_shot_elementwise(op, inputs, out, ctx)
-    } else {
-        fallback(inputs, out)
-    }
+    validate_read_into_destination(op.label(), inputs, &out)?;
+    let result = match op {
+        ElementwiseReadOp::Add => backend.add_read(inputs[0].clone(), inputs[1].clone())?,
+        ElementwiseReadOp::Subtract => backend.sub_read(inputs[0].clone(), inputs[1].clone())?,
+        ElementwiseReadOp::Multiply => backend.mul_read(inputs[0].clone(), inputs[1].clone())?,
+        ElementwiseReadOp::Negate => backend.neg_read(inputs[0].clone())?,
+        ElementwiseReadOp::Conj => backend.conj_read(inputs[0].clone())?,
+        ElementwiseReadOp::Divide => backend.div_read(inputs[0].clone(), inputs[1].clone())?,
+    };
+    backend.copy_read_into(TensorRead::from_tensor(&result), out)
 }
 
 /// Elementwise tensor operations.
@@ -2028,41 +1735,36 @@ pub fn elementwise_read_into_with_context(
 pub trait TensorElementwise: TensorStructural {
     /// Execute an elementwise operation into caller-owned storage.
     ///
-    /// Backend implementations normally override this hook only to inject
-    /// their explicit execution context and buffer policy. The default uses a
-    /// serial host one-shot kernel and preserves the allocating fallback for
-    /// device storage, dtype promotion, and broadcasting.
+    /// Every backend must implement this hook with its explicit execution
+    /// context, placement checks, and output-storage policy. The hook is the
+    /// backend boundary for borrowed inputs and caller-owned output; it must not
+    /// silently transfer data between devices.
+    ///
+    /// # Examples
+    ///
+    /// ```rust
+    /// use tenferro_tensor::{ElementwiseReadOp, TensorElementwise, TensorRead, TensorWrite};
+    ///
+    /// fn run_into<B: TensorElementwise>(
+    ///     backend: &mut B,
+    ///     input: TensorRead<'_>,
+    ///     out: TensorWrite<'_>,
+    /// ) -> tenferro_tensor::Result<()> {
+    ///     backend.elementwise_read_into(ElementwiseReadOp::Conj, &[input], out)
+    /// }
+    /// ```
     ///
     /// # Errors
     ///
-    /// Returns [`crate::Error::Validation`] when `inputs` has the wrong arity,
-    /// tensor metadata is invalid, or the destination overlaps an input.
-    /// Returns [`crate::Error::BackendSource`] when the strided kernel rejects
-    /// an eligible host operation. Errors from the allocating backend fallback
-    /// are preserved unchanged.
+    /// Returns [`crate::Error::Validation`] for invalid arity, metadata, or
+    /// overlapping storage. Backend-specific placement, unsupported-operation,
+    /// and execution errors are returned unchanged by the implementation.
     fn elementwise_read_into(
         &mut self,
         op: ElementwiseReadOp,
         inputs: &[TensorRead<'_>],
         out: TensorWrite<'_>,
-    ) -> crate::Result<()> {
-        let ctx = ExecContext::serial();
-        elementwise_read_into_with_context(op, inputs, out, &ctx, |inputs, out| {
-            let result = match op {
-                ElementwiseReadOp::Add => self.add_read(inputs[0].clone(), inputs[1].clone())?,
-                ElementwiseReadOp::Subtract => {
-                    self.sub_read(inputs[0].clone(), inputs[1].clone())?
-                }
-                ElementwiseReadOp::Multiply => {
-                    self.mul_read(inputs[0].clone(), inputs[1].clone())?
-                }
-                ElementwiseReadOp::Negate => self.neg_read(inputs[0].clone())?,
-                ElementwiseReadOp::Conj => self.conj_read(inputs[0].clone())?,
-                ElementwiseReadOp::Divide => self.div_read(inputs[0].clone(), inputs[1].clone())?,
-            };
-            self.copy_read_into(TensorRead::from_tensor(&result), out)
-        })
-    }
+    ) -> crate::Result<()>;
 
     /// # Errors
     ///
