+++ docs/architecture/tenferro-crates.md
@@ -38,6 +38,8 @@ stack. `tenferro-xla` is a peer executor over compiled static programs, not a
 | `tenferro-tensor-core` | Rank/layout metadata, dtype tags, scalar trait, and host-only tensor adapters |
 | `tenferro-tensor` | Runtime `TypedTensor<T, R>`/`Tensor` values, typed views, backend traits, and backend-independent contracts |
 | `tenferro-cpu` | Public CPU backend, CPU execution sessions, CPU execution context, provider selection, thread policy, public resource-pool controls, and the CPU runtime-registration preparation/execution adapter |
+| `tenferro-cpu-basic` | Shared CPU buffer pool, full-overwrite destination guard, and host strided-storage adapters used by CPU kernel families |
+| `tenferro-cpu-fused` | Internal CPU runtime-DAG fused elementwise adapter; delegates traversal to the fused strided kernel |
 | `tenferro-gpu` | CubeCL/CUDA backend and GPU transfer helpers |
 | `tenferro-runtime` | Concrete tensor helpers, traced tensors, graph compilation/execution, immutable runtime snapshots, preparation SPI/cache substrate, runtime-owned compiled-graph execution, scheduled-graph staging, extension runtime registration, and extension cache storage |
 | `tenferro-xla` | Experimental StableHLO lowering and runtime-loaded PJRT plugin support for static-shaped traced programs |
@@ -46,7 +48,7 @@ stack. `tenferro-xla` is a peer executor over compiled static programs, not a
 | `tenferro-linalg` | Linear algebra traced APIs, eager helpers, extension runtime, and optional linalg AD rules |
 | `tenferro-fft` | FFT extension runtime and public concrete/traced FFT APIs |
 | `tenferro-core-ops` | Internal core primitive operation catalog used by graph, runtime, and backend dispatch |
-| `tenferro-internal-cpu-kernels` | Internal CPU elementwise kernels and typed buffer-pool implementation reused by `tenferro-cpu` |
+| `tenferro-internal-cpu-kernels` | Internal ordinary dtype-dispatch CPU kernels and pool-aware one-shot read-into replay |
 | `tenferro-internal-ops` | Graph op vocabulary and AD rule implementations |
 | `tenferro-internal-extension-macros` | Procedural macros for extension-op registration |
 
@@ -94,8 +96,14 @@ Layer 2: tenferro-tensor
          tenferro-cpu
          public CPU backend and execution sessions
 
+         tenferro-cpu-basic
+         shared CPU pool, overwrite guard and host strided adapters
+
          tenferro-internal-cpu-kernels
-         internal CPU elementwise kernels and buffer-pool implementation
+         ordinary dtype-dispatch and read-into kernels
+
+         tenferro-cpu-fused
+         runtime-DAG fused elementwise adapter
 
          tenferro-gpu
          CubeCL/CUDA backend and GPU transfer helpers
@@ -123,30 +131,35 @@ runtime-owned execution bridge. It is not a runtime-to-CPU dependency:
 
 ```text
 tenferro-tensor           -> tenferro-tensor-core, tenferro-core-ops
-tenferro-cpu              -> tenferro-tensor, tenferro-core-ops
-tenferro-cpu              -> tenferro-internal-cpu-kernels
-tenferro-cpu              -> tenferro-runtime
+tenferro-cpu-basic        -> tenferro-tensor, strided-basic
+tenferro-internal-cpu-kernels
+                           -> tenferro-tensor, tenferro-cpu-basic,
+                              strided-kernel
+tenferro-cpu-fused        -> tenferro-tensor, tenferro-cpu-basic,
+                              strided-basic, strided-fused
+tenferro-cpu              -> tenferro-tensor, tenferro-core-ops,
+                              tenferro-cpu-basic, tenferro-cpu-fused,
+                              tenferro-internal-cpu-kernels,
+                              tenferro-runtime
 tenferro-gpu              -> tenferro-tensor, tenferro-core-ops,
                               tenferro-cpu, tenferro-runtime
-tenferro-internal-cpu-kernels
-                           -> tenferro-tensor
-tenferro-internal-ops     -> tenferro-tensor, tenferro-core-ops,
+tenferro-internal-ops    -> tenferro-tensor, tenferro-core-ops,
                               tenferro-internal-extension-macros
-tenferro-runtime          -> tenferro-tensor, tenferro-core-ops,
+tenferro-runtime         -> tenferro-tensor, tenferro-core-ops,
                               tenferro-internal-ops,
                               tenferro-internal-extension-macros
-tenferro-xla              -> tenferro-runtime, tenferro-internal-ops,
+tenferro-xla             -> tenferro-runtime, tenferro-internal-ops,
                               tenferro-tensor
-tenferro-ad               -> tenferro-runtime, tenferro-internal-ops,
+tenferro-ad              -> tenferro-runtime, tenferro-internal-ops,
                               tenferro-tensor, tenferro-cpu
 
-tenferro-einsum           -> tenferro-runtime, tenferro-internal-ops,
+tenferro-einsum          -> tenferro-runtime, tenferro-internal-ops,
                               tenferro-tensor, tenferro-cpu,
                               tenferro-internal-extension-macros
-tenferro-linalg           -> tenferro-runtime, tenferro-internal-ops,
+tenferro-linalg         -> tenferro-runtime, tenferro-internal-ops,
                               tenferro-tensor, tenferro-cpu,
                               tenferro-internal-extension-macros
-tenferro-fft              -> tenferro-runtime, tenferro-internal-ops,
+tenferro-fft            -> tenferro-runtime, tenferro-internal-ops,
                               tenferro-tensor, tenferro-cpu,
                               tenferro-internal-extension-macros
 ```
@@ -180,14 +193,22 @@ Rules:
 - `tenferro-tensor-core` must not expose public `TypedTensor` aliases.
   Backend-capable typed tensors are owned by `tenferro-tensor`.
 - `tenferro-tensor` owns concrete runtime tensor values, arbitrary-stride typed
-  views, backend traits, and backend-independent contracts.
+  views, backend traits, and backend-independent contracts. Its
+  `TensorElementwise::elementwise_read_into` hook is required at backend
+  implementations; tensor itself does not instantiate a CPU strided kernel.
 - `tenferro-cpu` owns `CpuBackend`, `CpuContext`, CPU execution sessions,
   CPU provider selection, thread policy, public resource-pool controls, and
   its narrow runtime-registration adapters for preparation metadata and
   runtime-owned execution.
-- `tenferro-internal-cpu-kernels` owns internal CPU elementwise kernels and the
-  typed buffer-pool implementation reused by `tenferro-cpu`. It exists as a
-  release-build reuse boundary and is not a user-facing API surface.
+- `tenferro-cpu-basic` owns the shared CPU buffer pool, full-overwrite guard,
+  and host strided-storage adapters. It has no dependency on ordinary or fused
+  CPU kernel families.
+- `tenferro-internal-cpu-kernels` owns ordinary dtype-dispatch elementwise
+  kernels and pool-aware one-shot read-into replay. It depends on basic but does
+  not own fused execution.
+- `tenferro-cpu-fused` owns runtime-DAG fused elementwise execution and depends
+  on basic plus `strided-fused`; it does not depend on ordinary kernel code.
+  Both families receive the enclosing CPU context and pool.
 - `tenferro-gpu` owns GPU backend implementation and transfer helpers.
 - `tenferro-runtime` owns graph construction, compilation, execution,
   immutable runtime snapshots/reconfiguration, preparation SPI/cache substrate,
