+++ docs/guides/parallelism-and-caching.md
@@ -64,9 +64,9 @@ workers remain provider-owned and may fan out independently.
 
 | Operation family | Threading behavior |
 | --- | --- |
-| Elementwise and analytic ops | `strided-kernel` map/zip kernels and erased fused replay use the already-entered context's native policy. Rayon-capable `Inner` may use the selected executor; all other modes above are sequential. |
+| Elementwise and analytic ops | Shared typed paths use `strided-basic`, ordinary dtype dispatch uses `strided-kernel`, and erased fused replay uses `strided-fused`; all use the already-entered context's native policy. Rayon-capable `Inner` may use the selected executor; all other modes above are sequential. |
 | Reductions | sum/product erased replay and typed max/min reductions use the same selected native policy and never ambient Rayon. Upstream strided plan coverage determines whether a specific multi-axis reduction shape can actually partition work. |
-| View materialization, transpose/permute, broadcast, convert, and diagonal extraction | `strided-kernel` copy/map kernels use the same selected native policy; layout fallback and linalg input materialization are included. |
+| View materialization, transpose/permute, broadcast, convert, and diagonal extraction | Strided copy/map kernels use the same selected native policy; layout fallback and linalg input materialization are included. |
 | `dot_general` through `cpu-faer` | faer receives `Par::rayon(n)` only for `Inner` execution whose selected executor advertises Rayon and whose validated budget is greater than one; otherwise it receives `Par::Seq`. Downstream direct faer calls can obtain the same scoped value through `FaerParallelismExt::with_faer_parallelism`, without exposing tenferro's executor. |
 | GEMM and linalg through `cpu-blas` | Threading is owned by the linked BLAS/LAPACK provider, not Rayon. Configure the provider variables below. |
 | Supported `dot_general` contractions through an external TBLIS provider | The example provider clamps TBLIS to one thread per call; unsupported TBLIS shapes fall back to the compiled faer/BLAS provider in preferred mode. |
